\name{RCircos.defaultBasePerUnits}
\alias{RCircos.defaultBasePerUnits}
\docType{data}
\title{
The Default Base Pair Number of A Circos Unit
}
\description{
The default total number of base pairs a Circos unit represents. This is a read only parameter and the actual value in a RCircos session could be modified with get and reset methods for plot parameters.
}
\format{
  The format is:
 num 30000
}
\keyword{ data }
