\name{RCircos.defaultCharWidth}
\alias{RCircos.defaultCharWidth}
\docType{data}
\title{
Default Character With in Circos Unit
}
\description{
Default Circos units of a character on RCircos plot. This is the character size for best resolution and is a read only parameter.
}
\format{
  The format is:
 num 500
}
\keyword{ data }
