\name{RCircos.defaultChromPadding}
\alias{RCircos.defaultChromPadding}
\docType{data}
\title{
The Default Value of GAP in Circos (Chromosome) Unit between two Chromosome Ideogram 
}
\description{
The default total unit of a gap bewtween the ends of two neighbor chromosome ideograms. This is a read only parameter.
}
\format{
  The format is:
 num 300
}
\keyword{ data }
