\name{RCircos.defaultCircosUnits}
\alias{RCircos.defaultCircosUnits}
\docType{data}
\title{
Default Circos Units for RCircos
}
\description{
The default Circos units for RCircos. A Circos unit is a point which represents an inteval of genome length on the circumustance of a circle line. By default an unit represents a point which equal to the length of 30000 base pairs and length of human genome is used to calculate the total number of Circos units. This is a read only parameter and is for internal use to modify plot parameters for RCircso plot with non-human genomes.
}
\format{
  The format is:
 num 103190
}
\keyword{ data }
