\name{RCircos.defaultTextSize}
\alias{RCircos.defaultTextSize}
\docType{data}
\title{
Default Character Size of RCircos Plot
}
\description{
The default character size for text plot of RCircos in best resolution. This is a read only parameter and the actual value in a RCircos session could be modified with get and reset methods for plot parameters.
}
\format{
  The format is:
 num 0.4
}
\keyword{ data }
