\name{RCircos.Env}
\alias{RCircos.Env}
\docType{data}
\title{
RCircos Environment
}
\description{
RCircos.Env is the user environment to hold RCircos core components and each component in the environment can be viewed with get mothod.
}
\format{
  The format is:
<environment: 0x14f6c10> 
}
\keyword{data}
