\name{RCircos.Gene.Label.Data}
\alias{RCircos.Gene.Label.Data}
\docType{data}
\title{
Sample Data for Gene Labels 
}
\description{
RCircos.Gene.Label.Data contains genomic position information for 192 genes. Each row of the data contains genomic position and gene name for one gene in the order of chromosome name, start position, end position, and gene name.  
}
\usage{data(RCircos.Gene.Label.Data)}
\format{
  A data frame with 192 observations on the following 4 variables.
  \describe{
    \item{\code{Chromosome}}{a factor with levels \code{chr1} \code{chr10} \code{chr11} \code{chr12} \code{chr13} \code{chr14} \code{chr15} \code{chr16} \code{chr17} \code{chr18} \code{chr19} \code{chr2} \code{chr20} \code{chr21} \code{chr22} \code{chr3} \code{chr4} \code{chr5} \code{chr6} \code{chr7} \code{chr8} \code{chr9} \code{chrX}}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
    \item{\code{Gene}}{a factor with levels \code{ABL1} \code{ACN9} \code{ACVR1B} \code{AKT1} \code{ALK} \code{APC} \code{ARHGEF2} \code{ARID1A} \code{ASXL1} \code{ATM} \code{ATRX} \code{BAG3} \code{BAI3} \code{BAP1} \code{BCAN} \code{BCL2} \code{BRAF} \code{BRCA1} \code{BRCA2} \code{CA12} \code{CA9} \code{CALU} \code{CARD11} \code{CBL} \code{CCND1} \code{CD34} \code{CDC73} \code{CDH1} \code{CDK4} \code{CDKN2A} \code{CDX2} \code{CEACAM7} \code{CEBPA} \code{CES3} \code{CRLF2} \code{CSF1R} \code{CSF3} \code{CTNNA1} \code{CTNNB1} \code{CUL2} \code{CYLD} \code{CYP1A1} \code{DAXX} \code{DCC} \code{DES} \code{DIRAS1} \code{DIRC2} \code{DKK3} \code{DLD} \code{DMBT1} \code{DNMT3A} \code{DPYD} \code{EGF} \code{EGFR} \code{EGR3} \code{EIF4G2} \code{EML4} \code{ENO1} \code{ENO2} \code{ERBB2} \code{ERBB3} \code{ERBB4} \code{ERG} \code{EZH2} \code{F10} \code{FAM123B} \code{FBXW7} \code{FGFR1} \code{FGFR2} \code{FGFR3} \code{FH} \code{FHIT} \code{FLCN} \code{FLT3} \code{FOXL2} \code{FZR1} \code{GATA1} \code{GATA2} \code{GATA3} \code{GFAP} \code{GNA11} \code{GNAQ} \code{GNAS} \code{GRP} \code{GSTM1} \code{HIVEP3} \code{HNF1A} \code{HRAS} \code{IDH1} \code{IDH2} \code{IFNA1} \code{IL2} \code{ITGB5} \code{JAK1} \code{JAK2} \code{JAK3} \code{KDR} \code{KIT} \code{KRAS} \code{KRT20} \code{MAGI1} \code{MAP2K4} \code{MEN1} \code{MET} \code{MGMT} \code{MIB1} \code{MKI67} \code{MLH1} \code{MPL} \code{MSH2} \code{MSH6} \code{MUC1} \code{MUC17} \code{MUC2} \code{MUTYH} \code{MVP} \code{MYC} \code{MYD88} \code{NAT1} \code{NAT2} \code{NES} \code{NF1} \code{NF2} \code{NOTCH1} \code{NOTCH2} \code{NPM1} \code{NRAS} \code{PBRM1} \code{PCNA} \code{PDGFRA} \code{PDZD4} \code{PGR} \code{PHOX2B} \code{PIK3CA} \code{PMS2} \code{PPP2R1A} \code{PRCC} \code{PRKAR1A} \code{PTCH1} \code{PTEN} \code{PTGS2} \code{PTPN11} \code{RB1} \code{REEP5} \code{RET} \code{RNF139} \code{RNF2} \code{RPS15} \code{RUNX1} \code{SDHA} \code{SDHAF1} \code{SDHAF2} \code{SDHB} \code{SDHC} \code{SDHD} \code{SELT} \code{SETD2} \code{SLC38A1} \code{SLC6A2} \code{SLTM} \code{SMAD4} \code{SMARCA4} \code{SMARCB1} \code{SMO} \code{SMOX} \code{SMUG1} \code{SOCS1} \code{SRC} \code{SST} \code{STC1} \code{STK11} \code{SUFU} \code{SYP} \code{TCF7L2} \code{TET2} \code{TFE3} \code{TFEB} \code{TMEM97} \code{TNFAIP3} \code{TNFSF13} \code{TP53} \code{TPD52L2} \code{TPM4} \code{TSC1} \code{TSC2} \code{TSHR} \code{TYK2} \code{VHL} \code{VIM} \code{WT1} \code{XRCC1} \code{ZNF135}}
  }
}
\source{
Unpublished data. 
}
\keyword{datasets}

