\name{RCircos.Get.Gene.Name.Plot.Parameters}
\alias{RCircos.Get.Gene.Name.Plot.Parameters}
\title{
Calculate Parameters for Gene Name Plot
}
\description{
Calculate maximum number of gene names to be plotted for each chromosome based on plot RCircos plot parameters. RCircos core components must be initialized first.
}
\usage{
RCircos.Get.Gene.Name.Plot.Parameters()
}
\value{
A data frame with names, maximum number of labels could be plotted, start position, end position, and label width for each chromosome.
}
\author{
Henry Zhang
}
\keyword{ methods }

