\name{RCircos.Get.Plot.Ideogram}
\alias{RCircos.Get.Plot.Ideogram}
\title{
Get Chromosome Ideogram Object
}
\description{
Get the chromosome ideogram object used during current RCircos session. RCircos core components must be initialized first.  This function is mainly for internal use. 
}
\usage{
RCircos.Get.Plot.Ideogram()
}
\value{
A data frame containing all information for chromosome ideogram plot. The first five column are orginal cytoband data and following are  chromosome highlight colors, and plot colors, length, number of chromosome unit, and relative location on the circular layout of each band.
}
\author{
Hongen Zhang
}
\keyword{ methods }

