\name{RCircos.heatmapColors}
\alias{RCircos.heatmapColors}
\docType{data}
\title{
Heatmap Colors Supported by Current Version of RCircos
}
\description{
Default heatmap colors (total of 6) supported by current version of RCircos. This is a read only data and can be viwed with the method of RCircos.Get.Supported.HeatmapColors().
}
\format{
  The format is:
 chr [1:6] "BlueWhiteRed" "GreenWhiteRed" "GreenYellowRed" 
"GreenBlackRed" "YellowToRed" "BlackOnly"
}
\keyword{datasets}
