\name{RCircos.Initialize.Parameters}
\alias{RCircos.Initialize.Parameters}
\title{
Initialize RCircos Plot Parameters
}
\description{
Setup default values for RCircos plot. All parameters are held in one list. This function is for internal use.
}
\usage{
RCircos.Initialize.Parameters(tracks.inside, tracks.outside)
}
\arguments{
  \item{tracks.inside}{
Integer, how many tracks will be put inside of chromosome ideogram.
}
  \item{tracks.outside}{
Integer, how many tracks will be put outside of chromosome ideogram.
}
}
\author{
Hongen Zhang
}
\keyword{ methods }
