\name{RCircos.Initialize.Plot.Parameters}
\alias{RCircos.Initialize.Plot.Parameters}
\title{
Initialize RCircos Plot Parameters
}
\description{
Setup default values for RCircos plot parameters. All parameters are held in one list. This function is for internal use only.
}
\usage{
RCircos.Initialize.Plot.Parameters(tracks.inside, tracks.outside)
}
\arguments{
  \item{tracks.inside}{
Non-negative integer, how many tracks will be put inside of chromosome ideogram.
}
  \item{tracks.outside}{
Non-negative integer, how many tracks will be put outside of chromosome ideogram.
}
}
\author{
Henry Zhang
}
\keyword{ methods }

