\name{RCircos.Line.Data}
\alias{RCircos.Line.Data}
\docType{data}
\title{
Sample Data for RCircos Line Plot 
}
\description{
A data frame with the first three columns as chromosome name, start and end position followed by columns of DNA copy number variant data (num.mark and seg.mean).  
}
\usage{data(RCircos.Line.Data)}
\format{
  A data frame with 2037 observations on the following 5 variables.
  \describe{
    \item{\code{chromosome}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{21} \code{22} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{X} \code{Y}}
    \item{\code{start}}{a numeric vector}
    \item{\code{stop}}{a numeric vector}
    \item{\code{num.mark}}{a numeric vector}
    \item{\code{seg.mean}}{a numeric vector}
  }
}
\source{
Unpublished data. 
}
\keyword{datasets}

