\name{RCircos.Link.Line}
\alias{RCircos.Link.Line}
\title{
Calculate x and y Coordinates for a Bezire Curve
}
\description{
Calculate a quandratic Bezire curve between two points with (0,0) as control points. This function is for internal use.
}
\usage{
RCircos.Link.Line(line.start=NULL, line.end=NULL)
}
\arguments{
  \item{line.start}{
The x and y coordinates of a point where the Bezier curve starts.
}
  \item{line.end}{
The x and y coordinates of a point where the Bezier curve ends.
}
}
\author{
Hongen Zhang
}
\keyword{ methods }

