\name{RCircos.paddingConst}
\alias{RCircos.paddingConst}
\docType{data}
\title{
Padding Distance Constant between Two Chromosomes
}
\description{
Default length of padding area between two chromosomes on chromosome ideogram track. This is a read only data and can be viewed with RCircos.Get.Padding.Constant() method.
}
\format{
  The format is:
 num 0.00291
}
\keyword{datasets}
