\name{RCircos.plotTypes}
\alias{RCircos.plotTypes}
\docType{data}
\title{
Plot Types Supported by Current RCircos Version
}
\description{
All supported plot types saved in RCircos Environment. The values are read only throug RCircos.Get.Supported.Plot.Types().
}
\format{
  The format is:
char[1:15] "link", "ribbon", "pLink", "polygon", "tile",  "ideogram", "heatmap", "bar", "histogram", "cLine","vLine", "points",  "connector", "ticks",  "text"
}
\keyword{datasets}
