\name{RCircos.Polygon.Data}
\alias{RCircos.Polygon.Data}
\docType{data}
\title{
RCircos Polygon Demo Data
}
\description{
A data frame with four columns for chromosome names, start and end position for each genomic interval, and values for polygon heights.
}
\usage{data("RCircos.Polygon.Data")}
\format{
  A data frame with 77 observations on the following 4 variables.
  \describe{
    \item{\code{Chromosome}}{a factor with levels \code{chr1} \code{chr10} \code{chr11} \code{chr12} \code{chr13} \code{chr14} \code{chr15} \code{chr16} \code{chr17} \code{chr18} \code{chr19} \code{chr2} \code{chr20} \code{chr21} \code{chr3} \code{chr4} \code{chr5} \code{chr6} \code{chr7} \code{chr8} \code{chr9} \code{chrX} \code{chrY}}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
    \item{\code{Data}}{a numeric vector}
  }
}
\keyword{datasets}
