\name{RCircos.Reset.Plot.Ideogram}
\alias{RCircos.Reset.Plot.Ideogram}
\title{
Reset RCircos Ideogram Plot Data
}
\description{
Reset chromosome ideogram plot data. Reserved for advanced usage.
}
\usage{
RCircos.Reset.Plot.Ideogram(chromIdeo)
}
\arguments{
  \item{chromIdeo}{
An object obtained from RCircos.Get.Plot.Ideogram() with/without modifications.
}
}
\author{
Hongen Zhang
}
\keyword{ methods }
