\name{RCircos.Tile.Data}
\alias{RCircos.Tile.Data}
\docType{data}
\title{
Sample Data for RCircos Tile Plot 
}
\description{
A data frame contining genomic position data only (chromosome, start and end positions). 
}
\usage{data(RCircos.Tile.Data)}
\format{
  A data frame with 152 observations on the following 3 variables.
  \describe{
    \item{\code{Chromosome}}{a factor with levels \code{chr1} \code{chr10} \code{chr11} \code{chr12} \code{chr13} \code{chr14} \code{chr15} \code{chr16} \code{chr17} \code{chr18} \code{chr19} \code{chr2} \code{chr20} \code{chr21} \code{chr22} \code{chr3} \code{chr4} \code{chr5} \code{chr6} \code{chr7} \code{chr8} \code{chr9} \code{chrX} \code{chrY}}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
  }
}
\source{
Unpublished Data. 
}
\keyword{datasets}

