\name{RCircos.Track.Positions}
\alias{RCircos.Track.Positions}
\title{
Calculate Data Plot Track Positions
}
\description{
Calculate inner and outer plot positions for a data track. RCircos core components must be initialized first. This function is for internal use only.
}
\usage{
RCircos.Track.Positions(side, track.num)
}
\arguments{
  \item{side}{
Character vector, either "in" or "out", representing the position related to chromosome ideogram.
}
  \item{track.num}{
Integer, representing the ordinal number of the plot track
}
}
\author{
Hongen Zhang
}
\keyword{ methods }

