\name{RCircos.Validate.Plot.Parameters}
\alias{RCircos.Validate.Plot.Parameters}
\title{
Validate Plot Parameters
}
\description{
Validate values of plot parameters before reset them in RCircos Environment. This function is mainly for internal use.
}
\usage{
RCircos.Validate.Plot.Parameters(parameters=NULL)
}
\arguments{
  \item{parameters}{
A list returned from RCircos.Get.Plot.Parameters().
}
}
\author{
Henry Zhang
}
\keyword{ methods }
