\name{RCircos.Workflow}
\alias{RCircos.Workflow}
\title{
Print Out RCircos Workflow 
}
\description{
Display  RCircos plot workflow on stardard output device for a quick quide.
}
\usage{
RCircos.Workflow()
}
\author{
Hongen Zhang
}
\examples{
library(RCircos);
RCircos.Workflow();
}
\keyword{ methods }

