\name{UCSC.Baylor.3.4.Rat.cytoBandIdeogram}
\alias{UCSC.Baylor.3.4.Rat.cytoBandIdeogram}
\docType{data}
\title{
Cytoband Information for Rat Chromosome Ideogram 
}
\description{
A data frame containing chromosome name, start and end position, band name, and intensity of Giemsa stains for each cytoband.  
}
\usage{data(UCSC.Baylor.3.4.Rat.cytoBandIdeogram)}
\format{
  A data frame with 246 observations on the following 5 variables.
  \describe{
    \item{\code{Chromosome}}{a factor with levels \code{chr1} \code{chr10} \code{chr11} \code{chr12} \code{chr13} \code{chr14} \code{chr15} \code{chr16} \code{chr17} \code{chr18} \code{chr19} \code{chr2} \code{chr20} \code{chr3} \code{chr4} \code{chr5} \code{chr6} \code{chr7} \code{chr8} \code{chr9} \code{chrX}}
    \item{\code{ChromStart}}{a numeric vector}
    \item{\code{ChromEnd}}{a numeric vector}
    \item{\code{Band}}{a factor with levels \code{p11} \code{p12} \code{p13} \code{p14} \code{p15} \code{p16} \code{p21} \code{p22} \code{q11} \code{q12} \code{q12.1} \code{q12.2} \code{q12.3} \code{q12.4} \code{q12.5} \code{q13} \code{q14} \code{q15} \code{q16} \code{q21} \code{q22} \code{q23} \code{q24} \code{q25} \code{q26} \code{q27} \code{q31} \code{q32} \code{q32.1} \code{q32.2} \code{q32.3} \code{q33} \code{q34} \code{q35} \code{q36} \code{q37} \code{q38} \code{q41} \code{q42} \code{q43} \code{q44} \code{q45} \code{q51} \code{q52} \code{q53} \code{q54} \code{q55}}
    \item{\code{Stain}}{a factor with levels \code{gneg} \code{gpos} \code{gvar}}
  }
}
\source{
http://genome.ucsc.edu 
}
\keyword{datasets}

