\name{UCSC.Mouse.GRCm38.CytoBandIdeogram}
\alias{UCSC.Mouse.GRCm38.CytoBandIdeogram}
\docType{data}
\title{
Cytoband Inforamtion for Mouse Chromosome Ideogram 
}
\description{
A data frame containing chromosome name, start and end position, band name, and intensity of Giemsa stains for each cytoband. 
}
\usage{data(UCSC.Mouse.GRCm38.CytoBandIdeogram)}
\format{
  A data frame with 403 observations on the following 5 variables.
  \describe{
    \item{\code{Chromosome}}{a factor with levels \code{chr1} \code{chr10} \code{chr11} \code{chr12} \code{chr13} \code{chr14} \code{chr15} \code{chr16} \code{chr17} \code{chr18} \code{chr19} \code{chr2} \code{chr3} \code{chr4} \code{chr5} \code{chr6} \code{chr7} \code{chr8} \code{chr9} \code{chrX} \code{chrY}}
    \item{\code{ChromStart}}{a numeric vector}
    \item{\code{ChromEnd}}{a numeric vector}
    \item{\code{Band}}{a factor with levels \code{qA} \code{qA1} \code{qA1.1} \code{qA1.2} \code{qA1.3} \code{qA2} \code{qA3} \code{qA3.1} \code{qA3.2} \code{qA3.3} \code{qA4} \code{qA5} \code{qA5.1} \code{qA5.2} \code{qA5.3} \code{qA6} \code{qA7.1} \code{qA7.2} \code{qA7.3} \code{qB} \code{qB1} \code{qB1.1} \code{qB1.2} \code{qB1.3} \code{qB2} \code{qB2.1} \code{qB2.2} \code{qB2.3} \code{qB3} \code{qB3.1} \code{qB3.2} \code{qB3.3} \code{qB4} \code{qB5} \code{qB5.1} \code{qB5.2} \code{qB5.3} \code{qC} \code{qC1} \code{qC1.1} \code{qC1.2} \code{qC1.3} \code{qC2} \code{qC3} \code{qC3.1} \code{qC3.2} \code{qC3.3} \code{qC4} \code{qC5} \code{qC6} \code{qC7} \code{qD} \code{qD1} \code{qD2} \code{qD2.1} \code{qD2.2} \code{qD2.3} \code{qD3} \code{qE} \code{qE1} \code{qE1.1} \code{qE1.2} \code{qE1.3} \code{qE2} \code{qE2.1} \code{qE2.2} \code{qE2.3} \code{qE3} \code{qE3.1} \code{qE3.2} \code{qE3.3} \code{qE4} \code{qE5} \code{qF} \code{qF1} \code{qF2} \code{qF2.1} \code{qF2.2} \code{qF2.3} \code{qF3} \code{qF4} \code{qF5} \code{qG1} \code{qG1.1} \code{qG1.2} \code{qG1.3} \code{qG2} \code{qG3} \code{qH1} \code{qH2} \code{qH2.1} \code{qH2.2} \code{qH2.3} \code{qH3} \code{qH4} \code{qH5} \code{qH6}}
    \item{\code{Stain}}{a factor with levels \code{gneg} \code{gpos100} \code{gpos33} \code{gpos66} \code{gpos75}}
  }
}
\source{
http://genome.ucsc.edu 
}
\keyword{datasets}

