\name{RCircos-package}
\alias{RCircos-package}
\alias{RCircos}
\docType{package}
\title{
Circos-like Plot of Genomic Data
}
\description{
This is the R implement of basic Circos plots of genimic data for: chromosome ideogram, connectors, gene labels, heatmap, histogram, line, links, scatter, and tiles.
}
\details{
\tabular{ll}{
Package: \tab RCircos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-11\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Hongen Zhang 

Maintainer: Hongen Zhang <hzhang@mail.nih.gov> 
}
\references{
Hongen Zhang, Paul Meltzer, and Sean Davis. RCircos: an R package for Circos 2D track plot of geneomic data. To be published soon.
}
\keyword{ package }
\examples{
library(RCircos) 
}
