\name{RCircos.Env}
\alias{RCircos.Env}
\docType{data}
\title{
RCircos Environment
}
\description{
RCircos.Env is the user environment to hold RCircos core components and other objects for RCircos plot. Each object in the environment can be viewed with get mothod.
}
\format{
  The format is:
<environment: 0xd67930> 
}
\keyword{ data }
