\name{RCircos.Link.Line}
\alias{RCircos.Link.Line}
\title{
Draw A Circos Link Line
}
\description{
Calculate and draw a quandratic Bezire curve between two points with (0,0) as control points. This function is for internal use.
}
\usage{
RCircos.Link.Line(line.start, line.end, line.color)
}
\arguments{
  \item{line.start}{
The position (a point) where the Bezier curve starts.
}
  \item{line.end}{
The position (a point) where the Bezier curve ends.
}
  \item{line.color}{
Color for the Bezier curve. Any value that represents a R color.
}
}
\author{
Hongen Zhang
}
\keyword{ methods }
