\name{RCircos.Link.Line}
\alias{RCircos.Link.Line}
\title{
Calculate x and y Coordinates for a Bezire Curve
}
\description{
Calculate a quandratic Bezire curve between two points with (0,0) as control points. This function is for internal use.
}
\usage{
RCircos.Link.Line(line.start, line.end)
}
\arguments{
  \item{line.start}{
The position (a point) where the Bezier curve starts.
}
  \item{line.end}{
The position (a point) where the Bezier curve ends.
}
}
\author{
Hongen Zhang
}
\keyword{ methods }
