\name{RCircos.List.Parameters}
\alias{RCircos.List.Parameters}
\title{
Print out Graphic Parameters to Standard Output
}
\description{
Display all graphic parameters specific for Circos plot on standard output.
}
\usage{
RCircos.List.Parameters(RCircos.Par)
}
\arguments{
  \item{RCircos.Par}{
A list returned from RCircos.Initialize.Parameters() that contains graphics parameters specific for Circos plot.
}
}
\author{
Hongen Zhang
}
\keyword{methods}
