\name{RCircos.Workflow}
\alias{RCircos.Workflow}
\title{
A Simple RCircos Plot Workflow
}
\description{
Print out on screen to list procedures for a basic RCircos plot.
}
\usage{
RCircos.Workflow()
}
\author{
Henry Zhang
}
\examples{
\dontrun{
library(RCircos)
RCircos.Workflow()
}
}
\keyword{ methods }

